% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_tp_6b.R
\name{fn_tp_6b}
\alias{fn_tp_6b}
\title{Transaction Profile Function to process 6 bolts}
\usage{
fn_tp_6b(df_tp_6b_in)
}
\arguments{
\item{df_tp_6b_in}{Do you love cats? Defaults to TRUE.}
}
\description{
Transaction Profile Function to process 6 bolts
}
\examples{
df_tp_6b_in <- rapi.tp::fn_tp_6b(df_tp_6b_in)
}
\keyword{Transaction}
